Adjust weapon byte flag data types for TR4/TR5

+++ TRR-SaveMaster/TR4Utilities.cs
@@ -44,8 +44,8 @@ namespace TRR_SaveMaster
         private const int CROSSBOW_EXPLOSIVE_AMMO_OFFSET = 0x1DA;
 
         // Weapon byte flags
-        private const int WEAPON_PRESENT = 0x9;
-        private const int WEAPON_PRESENT_WITH_SIGHT = 0xD;
+        private const byte WEAPON_PRESENT = 0x9;
+        private const byte WEAPON_PRESENT_WITH_SIGHT = 0xD;
 
         // Health
         private const UInt16 MAX_HEALTH_VALUE = 1000;
